Update on "[quant][pt2e] Add early prototype top level quantize_pt2e APIs"
Summary:
This PR introduces the top level APIs for quantization support in PyTorch 2.0 Export stack
* prepare_pt2e
Takes a model that is captured by the PyTorch 2.0 export (torchdynamo full graph mode) and prepares the model for calibration
for post training quantization

* convert_pt2e
Takes a calibrated model and converts that to a reference quantized model that can be lowered later to quantized operator libraries or delegation modules

Note: everything related to quantize_pt2 are experimental (prototype), and we don't have any bc guarantees

Test Plan:
python test/test_quantization.py TestQuantizePT2EModels

Reviewers:

Subscribers:

Tasks:

Tags:

[ghstack-poisoned]
